Refactor layouts 1 (#1402)
* Rename Layout to SwapLayout

* Move SwapLayout to it's own namespace

* Rename exported component name

* Export SwapLayout

* Move PageWrapper out of SwapLayout

* Rename wrappers

* Simplify import
