--- src/components/Layout/Card/index.tsx
@@ -1,1 +1,0 @@
-export * from 'components/Layout/Card/Card'
