+++ src/components/layout/SwapLayout/Header/useNavigation.ts
@@ -1,8 +1,8 @@
 import { useEffect, useState } from 'react'
 import { useLocation } from 'react-router'
 
-import { HeaderProps } from 'components/Layout/Header'
-import { HeaderNavLinksInterface } from 'components/Layout/Header/Navigation'
+import { HeaderProps } from 'components/layout/SwapLayout/Header'
+import { HeaderNavLinksInterface } from 'components/layout/SwapLayout/Header/Navigation'
 
 const useNavigation = (initialState: HeaderProps['navigation'], responsive: boolean): HeaderNavLinksInterface[] => {
   const [navigationArray, setNavigationArray] = useState(initialState)
