+++ src/components/layout/index.ts
@@ -1,0 +1,2 @@
+export { SwapLayout } from './SwapLayout/index'
+export * from './PageWrappers'
